Function load_balancer::main

pub(crate) fn main() -> Result<()>

Point d’entrée principal de l’application. Configure le load balancer et écoute les connexions entrantes.

Cette fonction utilise Tokio pour gérer des opérations asynchrones, notamment l’écoute de connexions TCP, la gestion d’un cache partagé et la redirection des connexions vers des serveurs cibles.

Returns

tokio::io::Result<()> - Un résultat indiquant le succès ou l’échec de l’exécution de la fonction.

Examples

#[tokio::main]
async fn main() -> tokio::io::Result<()> {
    let listener = TcpListener::bind("127.0.0.1:7878").await?;
    println!("Load balancer running on localhost:7878");

    let cache = Arc::new(Mutex::new(Cache::new()));

    loop {
        let (mut socket, addr) = listener.accept().await?;
        let cache = Arc::clone(&cache);

        tokio::spawn(async move {
            let ip = addr.ip().to_string();
            let mut cache = cache.lock().await;
            let server = cache.get_server(&ip).await;

            let now = SystemTime::now();
            println!("Redirecting connection from: {} to {} at {:?}", ip, server, now);

            let mut server_socket = TcpStream::connect(server).await.unwrap();
            let mut buf = [0; 1024];

            let n = socket.read(&mut buf).await.unwrap();
            if n == 0 { return; }

            if server_socket.write_all(&buf[..n]).await.is_err() {
                eprintln!("Failed to write to server");
                return;
            }

            match server_socket.read(&mut buf).await {
                Ok(0) => return,
                Ok(n) => {
                    if socket.write_all(&buf[..n]).await.is_err() {
                        eprintln!("Failed to write back to client");
                        return;
                    }
                },
                Err(e) => {
                    eprintln!("Failed to read from server: {}", e);
                    return;
                }
            }
        });
    }
}

Panics

Cette fonction ne devrait pas paniquer dans des conditions normales d’utilisation.

Errors

Cette fonction retourne une erreur de type tokio::io::Error si elle échoue à lier le listener TCP ou à accepter une connexion.

Tokio

Cette fonction utilise l’attribut #[tokio::main] pour indiquer qu’elle est le point d’entrée d’une application Tokio asynchrone.