Module load_balancer::response

Main Module

This module contains the main entry point for the load balancer application. It initializes the necessary components and starts the load balancer server.

Usage

To run the application, use the following command:

cargo run -- --load-balancer-ip 192.168.1.196:5555 -b 192.168.1.192:2222,192.168.1.190:8080,192.168.1.191:3333 -p /health-check -i 3 -s 10 -r 5

For the web servers, we used the actix-web crate to create a simple web server that listens on the specified port and has paths for the health check and the / endpoint. These web servers are hosted on LCX containers created on a Proxmox server. The load balancer test machine is hosted on a separate LCX container on the same Proxmox server.

Re-exports

pub use gen::generator;

Modules

gen 🔒

This module provides utilities for generating HTTP responses.