Module load_balancer::response::gen

This module provides utilities for generating HTTP responses.

It includes the generator function, which can be used to generate HTTP responses with specified status codes, content types, and content.

Examples

use crate::gen::generator;

let response = generator("200 OK", "text/plain", "Hello, World!");
println!("Generated Response: {}", response);

The gen module contains the generator function, and it can be accessed using use crate::gen::generator;.

Functions

generator

Generates an HTTP response with the specified status code, content type, and content.